requirements.txt
opencv-python==4.12.0.88
PyQt5==5.15.9
pyzbar==0.1.9
Pillow==10.0.1
mariadb==1.1.8
numpy==2.2.6
pandas==2.3.1
matplotlib==3.10.5